phase 4

+++ SRC/emailManager/emails/views.py
@@ -1,4 +1,6 @@
+import datetime
 import email.errors
+
 from django.contrib.postgres.search import SearchVector
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db.models import Q
@@ -16,6 +18,7 @@ from emails.models import *
 from django.contrib.postgres.search import TrigramSimilarity
 from django.db.models.functions import Greatest
 from users.models import Signature
+
 
 class AddNewEmail(LoginRequiredMixin, View):
     template_name = 'emails/compose.html'
@@ -285,7 +288,22 @@ class EmailInboxView(LoginRequiredMixin, View):
     def get(self, request):
         place_holder = EmailPlaceHolders.objects.get(place_holder='inbox')
         user_inbox = UserEmailMapped.objects.filter(user=self.request.user, place_holder=place_holder)
-        # label = EmailPlaceHolders.objects.filter(creator=request.user)
+        for mail in user_inbox:
+
+            # print("*%"*79)
+            # print(filtering)
+            if Filter.objects.filter(user=request.user, subject=mail.email.subject, sender=mail.email.author).exists():
+                filtering = Filter.objects.filter(user=request.user, subject=mail.email.subject,
+                                                  sender=mail.email.author)
+                filtering.update(email=mail.email)
+                for filters in filtering:
+                    # user_inbox.update(place_holder=filters.place_holder)
+                    if mail.email == filters.email:
+                        print("*%" * 79)
+                        print(mail.email == filters.email)
+                        user_inbox = user_inbox.filter(email=filters.email)
+                        user_inbox.update(place_holder=filters.place_holder)
+
         return render(request, self.template_name, {'inbox': user_inbox})
 
 
@@ -323,7 +341,9 @@ class DetailEmailView(LoginRequiredMixin, View):
                 has_bcc = True
         except:
             bcc = ''
-
+        inbox = EmailPlaceHolders.objects.get(place_holder="inbox")
+        email_mapp = UserEmailMapped.objects.filter(email=email, place_holder=inbox)
+        email_mapp.update(is_read=True)
         return render(request, self.template, {'object': email, "to": to, "cc": cc, "bcc": bcc, "has_bcc": has_bcc})
 
 
@@ -366,6 +386,13 @@ class EmailTrashView(LoginRequiredMixin, View):
     def get(self, request):
         place_holder = EmailPlaceHolders.objects.get(place_holder='trash')
         user_trash = UserEmailMapped.objects.filter(user=self.request.user, place_holder=place_holder)
+        for emails in user_trash:
+            days = emails.email.created_date + datetime.timedelta(days=30)
+            user_trash_02 = UserEmailMapped.objects.filter(user=self.request.user, place_holder=place_holder,
+                                                           email__created_date__gte=days)
+            if user_trash_02.exists():
+                for email_trash in user_trash_02:
+                    email_trash.delete()
         label = EmailPlaceHolders.objects.filter(creator=request.user)
         return render(request, self.template_name, {'trash': user_trash, "label": label})
 
@@ -405,9 +432,22 @@ class DeleteLabel(LoginRequiredMixin, View):
         label = EmailPlaceHolders.objects.get(pk=label_id)
         emails = UserEmailMapped.objects.filter(place_holder=label)
         inbox = EmailPlaceHolders.objects.get(place_holder='inbox')
-        for email in emails:
-            user_email_mapped = UserEmailMapped.objects.filter(place_holder=label, user=request.user, email=email.email)
-            user_email_mapped.delete()
+        if emails.exists():
+            for email in emails:
+                user_email_mapped = UserEmailMapped.objects.filter(place_holder=label, user=request.user,
+                                                                   email=email.email)
+                user_email_mapped.delete()
+                if Filter.objects.filter(user=request.user, email=email.email, place_holder=label).exists():
+                    filtering = Filter.objects.filter(user=request.user, email=email.email, place_holder=label)
+                    print("_+" * 90)
+                    print(filtering)
+                    user_email_mapped_label = UserEmailMapped(email=email.email, user=request.user,
+                                                              place_holder=inbox)
+                    user_email_mapped_label.save()
+                    filtering.delete()
+
+                    label.delete()
+        else:
             label.delete()
         messages.success(request, "delete was successfully", 'success')
         return redirect('email_view')
@@ -429,15 +469,28 @@ class EmailArchiveView(LoginRequiredMixin, View):
     def get(self, request):
         place_holder = EmailPlaceHolders.objects.get(place_holder='archive')
         email = UserEmailMapped.objects.filter(place_holder=place_holder, user=request.user)
+        for mail in email:
+            if Filter.objects.filter(user=request.user, subject=mail.email.subject, sender=mail.email.author).exists():
+                filtering = Filter.objects.filter(user=request.user, subject=mail.email.subject,
+                                                  sender=mail.email.author)
+                filtering.update(email=mail.email)
+                for filters in filtering:
+                    # user_inbox.update(place_holder=filters.place_holder)
+                    if mail.email == filters.email:
+                        print("*%" * 79)
+                        print(mail.email == filters.email)
+                        email = email.filter(email=filters.email)
+                        email.update(place_holder=filters.place_holder)
+
         return render(request, self.template_name, {"email": email})
 
 
 class Search(LoginRequiredMixin, View):
     def get(self, request):
         user_email_mapped = UserEmailMapped.objects.filter(user=request.user)
-        print('^' * 60)
-        print(request.user)
-        print('^' * 60)
+        # print('^' * 60)
+        # print(request.user)
+        # print('^' * 60)
         form = SearchBox()
         final_query = Q()
         if 'search' in request.GET:
@@ -464,4 +517,19 @@ class Search(LoginRequiredMixin, View):
 
 class Setting(LoginRequiredMixin, View):
     def get(self, request):
+        return render(request, 'emails/setting.html')
+
+
+class AddLabelFilter(LoginRequiredMixin, View):
+    def get(self, request):
+        return render(request, 'emails/add_label_filter.html')
+
+    def post(self, request):
+        place_holder = request.POST['place_holder']
+        sender = request.POST['sender']
+        subject = request.POST['subject']
+        label = EmailPlaceHolders(place_holder=place_holder, creator=request.user)
+        label.save()
+        add_filter = Filter(place_holder=label, sender=sender, subject=subject, user=request.user)
+        add_filter.save()
         return render(request, 'emails/setting.html')
